Fix addons source for 2021.01
It's 10:00 and I'm a couple beers in, so third time's the charm as I try to get SOURCE_URL and SOURCE_SUM matching for both this file and `app.src`.

I should have done this on a new branch but WHOOPSIE-DOODLE

Seriously, y'all, I'm not deliberately trying to illustrate this, but I should not be trusted as the maintainer for this repo. I'm not the schmuck you want doing this.