Update systemd.service
This seems to be fixing a new issue with warning: mail_queue_enter: create file maildrop/513692.12834: Permission denied

Posts posted on friendica failed to be delivered. I have used chatgpt (need to be honest) but also did research on my own. I tried many things. In the end these changes to the systemd file worked. 

I ask if anyone can check it please to make sure it is fine.